Remove extra stuff (#3)
Remove obviously superfluous files.
1)  [Catwatch](https://github.com/zalando-incubator/catwatch) - seems to be a Zalando-specific tool.
2) [Zappr](https://zappr.readthedocs.io/en/latest/) - Some sort of review framework predating GitHub native controls.
3) `/legacy` directory - An artifact of [migration from Markdown to Asciidoc](https://github.com/zalando/restful-api-guidelines/pull/282) (apparently hosted on some website somewhere) circa 2017.
4) `/models` directory - The style guide references the hosted Zalando version, which is still valid.
We use common-schemas and style guide should be updated to point there where
applicable as it is retrofitted.

I didn't remove .travis.yaml or other build related things. I will do that when I replace with a Codefresh pipeline(s). This is also more of a "axe" removal than a "pocket knife" one. I'm sure we will remove other little bits and pieces from files as we dig in. This is just chopping away the obvious cruft to get closer to the general shape before more detailed carving.